requirements.txt
aiohttp~=3.12.0
certifi~=2025.4.26
python-dotenv~=1.1.0
aiomysql~=0.2.0
requests~=2.32.3
PyMySQL~=1.1.1
